Moved border styling constants out of theme index file
